+++ lib/src/dialogs/dialogs.dart
@@ -2,4 +2,3 @@ export 'actions.dart';
 export 'basic_bottom_sheet.dart';
 export 'default_picker_dialog.dart';
 export 'platform_bottom_sheet.dart';
-export 'platform_widget.dart';