+++ lib/src/input_formatters/input_formatters.dart
@@ -1,4 +1,5 @@
 export 'case_text_formatters.dart';
+export 'decimal_text_formatter.dart';
 export 'emoji_text_formatter.dart';
 export 'masked_text_formatter.dart';
 export 'no_space_text_formatter.dart';